Move proxy to devtools
