+++ packages/core/agent/src/plugins/dashboard/index.ts
@@ -3,4 +3,3 @@
 //
 
 export * from './dashboard-plugin';
-export * from './dashboard-proxy';
